[math] Update pytests
