+++ brainpy/_src/math/jitconn/tests/test_matvec.py
@@ -4,27 +4,18 @@ from functools import partial
 import jax
 import jax.numpy as jnp
 from absl.testing import parameterized
+from .._matvec import (mv_prob_homo_brainpylib as brainpylib_mv_prob_homo,
+                       mv_prob_uniform_brainpylib as brainpylib_mv_prob_uniform,
+                       mv_prob_normal_brainpylib as brainpylib_mv_prob_normal,)
 
 import brainpy.math as bm
-import platform
-import pytest
 
-is_manual_test = False
-if platform.system() == 'Windows' and not is_manual_test:
-  pytest.skip('brainpy.math package may need manual tests.', allow_module_level=True)
+shapes = [(100, 200), (10, 1000), (2, 1000), (1000, 10), (1000, 2)]
+shapes = [(100, 200), (2, 1000), (1000, 2)]
 
-shapes = [(100, 200),
-          (10, 1000),
-          (2, 1000),
-          (1000, 10),
-          (1000, 2)]
-
-brainpylib_mv_prob_homo = partial(bm.jitconn.mv_prob_homo, method='brainpylib')
-taichi_mv_prob_homo = partial(bm.jitconn.mv_prob_homo, method='taichi')
-brainpylib_mv_prob_uniform = partial(bm.jitconn.mv_prob_uniform, method='brainpylib')
-taichi_mv_prob_uniform = partial(bm.jitconn.mv_prob_uniform, method='taichi')
-brainpylib_mv_prob_normal = partial(bm.jitconn.mv_prob_normal, method='brainpylib')
-taichi_mv_prob_normal = partial(bm.jitconn.mv_prob_normal, method='taichi')
+taichi_mv_prob_homo = bm.jitconn.mv_prob_homo
+taichi_mv_prob_uniform = bm.jitconn.mv_prob_uniform
+taichi_mv_prob_normal = bm.jitconn.mv_prob_normal
 
 class Test_matvec_prob_conn(parameterized.TestCase):
   def __init__(self, *args, platform='cpu', **kwargs):
@@ -66,34 +57,32 @@ class Test_matvec_prob_conn(parameterized.TestCase):
     rng = bm.random.RandomState()
     vector = bm.as_jax(rng.random(shape[0] if transpose else shape[1]))
 
-    r1 = brainpylib_mv_prob_homo(vector,
-                                 homo_data,
-                                 conn_prob=prob,
-                                 shape=shape,
-                                 seed=seed,
-                                 outdim_parallel=outdim_parallel,
-                                 transpose=transpose)
+    r1 = taichi_mv_prob_homo(vector,
+                                        homo_data,
+                                        conn_prob=prob,
+                                        shape=shape,
+                                        seed=seed,
+                                        outdim_parallel=outdim_parallel,
+                                        transpose=transpose)
 
-    r2 = brainpylib_mv_prob_homo(vector,
-                                 homo_data,
-                                 conn_prob=prob,
-                                 shape=shape,
-                                 seed=seed,
-                                 outdim_parallel=outdim_parallel,
-                                 transpose=transpose)
+    r2 = taichi_mv_prob_homo(vector,
+                                        homo_data,
+                                        conn_prob=prob,
+                                        shape=shape,
+                                        seed=seed,
+                                        outdim_parallel=outdim_parallel,
+                                        transpose=transpose)
     self.assertTrue(jnp.allclose(r1, r2))
 
-    r2 = brainpylib_mv_prob_homo(vector,
-                                 homo_data,
-                                 conn_prob=prob,
-                                 shape=(shape[1], shape[0]),
-                                 seed=seed,
-                                 outdim_parallel=outdim_parallel,
-                                 transpose=not transpose)
+    r2 = taichi_mv_prob_homo(vector,
+                                        homo_data,
+                                        conn_prob=prob,
+                                        shape=(shape[1], shape[0]),
+                                        seed=seed,
+                                        outdim_parallel=outdim_parallel,
+                                        transpose=not transpose)
     self.assertTrue(jnp.allclose(r1, r2))
 
-    if x64:
-      bm.disable_x64()
     bm.clear_buffer_memory()
 
   @parameterized.named_parameters(
@@ -128,11 +117,11 @@ class Test_matvec_prob_conn(parameterized.TestCase):
     weights = bm.as_jax(rng.random(10))
 
     f1 = jax.vmap(
-      lambda event, data: brainpylib_mv_prob_homo(
+      lambda event, data: taichi_mv_prob_homo(
         event, data,
         conn_prob=prob, shape=shape, seed=seed,
         outdim_parallel=outdim_parallel, transpose=transpose
-      )
+      )[0]
     )
     r1 = f1(events, weights)
     r2 = f1(events, weights)
@@ -173,14 +162,14 @@ class Test_matvec_prob_conn(parameterized.TestCase):
     events = events.astype(float)
 
     f1 = jax.grad(
-      lambda event, data: brainpylib_mv_prob_homo(
+      lambda event, data: taichi_mv_prob_homo(
         event, data,
         conn_prob=prob,
         shape=shape,
         seed=seed,
         outdim_parallel=outdim_parallel,
         transpose=transpose
-      ).sum(),
+      )[0].sum(),
       argnums=0
     )
     r1 = f1(events, 1.)
@@ -230,36 +219,36 @@ class Test_matvec_prob_conn(parameterized.TestCase):
     rng = bm.random.RandomState()
     events = bm.as_jax(rng.random(shape[0] if transpose else shape[1]))
 
-    r1 = brainpylib_mv_prob_uniform(events,
-                                    w_low=w_low,
-                                    w_high=w_high,
-                                    conn_prob=prob,
-                                    shape=shape,
-                                    seed=seed,
-                                    outdim_parallel=outdim_parallel,
-                                    transpose=transpose)
+    r1 = taichi_mv_prob_uniform(events,
+                                           w_low=w_low,
+                                           w_high=w_high,
+                                           conn_prob=prob,
+                                           shape=shape,
+                                           seed=seed,
+                                           outdim_parallel=outdim_parallel,
+                                           transpose=transpose)
 
-    r2 = brainpylib_mv_prob_uniform(events,
-                                    w_low=w_low,
-                                    w_high=w_high,
-                                    conn_prob=prob,
-                                    shape=shape,
-                                    seed=seed,
-                                    outdim_parallel=outdim_parallel,
-                                    transpose=transpose)
+    r2 = taichi_mv_prob_uniform(events,
+                                           w_low=w_low,
+                                           w_high=w_high,
+                                           conn_prob=prob,
+                                           shape=shape,
+                                           seed=seed,
+                                           outdim_parallel=outdim_parallel,
+                                           transpose=transpose)
     c = jnp.allclose(r1, r2)
     if not c:
       print(r1, r2)
     self.assertTrue(c)
 
-    r2 = brainpylib_mv_prob_uniform(events,
-                                    w_low=w_low,
-                                    w_high=w_high,
-                                    conn_prob=prob,
-                                    shape=(shape[1], shape[0]),
-                                    seed=seed,
-                                    outdim_parallel=outdim_parallel,
-                                    transpose=not transpose)
+    r2 = taichi_mv_prob_uniform(events,
+                                           w_low=w_low,
+                                           w_high=w_high,
+                                           conn_prob=prob,
+                                           shape=(shape[1], shape[0]),
+                                           seed=seed,
+                                           outdim_parallel=outdim_parallel,
+                                           transpose=not transpose)
     c = jnp.allclose(r1, r2)
     if not c:
       print(r1, r2)
@@ -298,14 +287,14 @@ class Test_matvec_prob_conn(parameterized.TestCase):
     rng = bm.random.RandomState()
     events = bm.as_jax(rng.random((10, shape[0] if transpose else shape[1])))
 
-    f1 = jax.vmap(lambda e: brainpylib_mv_prob_uniform(e,
-                                                       w_low=0.,
-                                                       w_high=1.,
-                                                       conn_prob=prob,
-                                                       shape=shape,
-                                                       seed=seed,
-                                                       outdim_parallel=outdim_parallel,
-                                                       transpose=transpose))
+    f1 = jax.vmap(lambda e: taichi_mv_prob_uniform(e,
+                                                              w_low=0.,
+                                                              w_high=1.,
+                                                              conn_prob=prob,
+                                                              shape=shape,
+                                                              seed=seed,
+                                                              outdim_parallel=outdim_parallel,
+                                                              transpose=transpose))
 
     r1 = f1(events)
     r2 = f1(events)
@@ -347,7 +336,7 @@ class Test_matvec_prob_conn(parameterized.TestCase):
     events = bm.as_jax(rng.random(shape[0] if transpose else shape[1]))
 
     f1 = jax.grad(
-      lambda e, w_low, w_high: brainpylib_mv_prob_uniform(
+      lambda e, w_low, w_high: taichi_mv_prob_uniform(
         e,
         w_low=w_low,
         w_high=w_high,
@@ -356,7 +345,7 @@ class Test_matvec_prob_conn(parameterized.TestCase):
         seed=seed,
         outdim_parallel=outdim_parallel,
         transpose=transpose
-      ).sum()
+      )[0].sum()
     )
 
     r1 = f1(events, 0., 1.)
@@ -407,36 +396,36 @@ class Test_matvec_prob_conn(parameterized.TestCase):
     rng = bm.random.RandomState()
     events = bm.as_jax(rng.random(shape[0] if transpose else shape[1]))
 
-    r1 = brainpylib_mv_prob_normal(events,
-                                   w_mu=w_mu,
-                                   w_sigma=w_sigma,
-                                   conn_prob=prob,
-                                   shape=shape,
-                                   seed=seed,
-                                   outdim_parallel=outdim_parallel,
-                                   transpose=transpose)
+    r1 = taichi_mv_prob_normal(events,
+                                          w_mu=w_mu,
+                                          w_sigma=w_sigma,
+                                          conn_prob=prob,
+                                          shape=shape,
+                                          seed=seed,
+                                          outdim_parallel=outdim_parallel,
+                                          transpose=transpose)
 
-    r2 = brainpylib_mv_prob_normal(events,
-                                   w_mu=w_mu,
-                                   w_sigma=w_sigma,
-                                   conn_prob=prob,
-                                   shape=shape,
-                                   seed=seed,
-                                   outdim_parallel=outdim_parallel,
-                                   transpose=transpose)
+    r2 = taichi_mv_prob_normal(events,
+                                          w_mu=w_mu,
+                                          w_sigma=w_sigma,
+                                          conn_prob=prob,
+                                          shape=shape,
+                                          seed=seed,
+                                          outdim_parallel=outdim_parallel,
+                                          transpose=transpose)
     c = jnp.allclose(r1, r2)
     if not c:
       print(r1, r2)
     self.assertTrue(c)
 
-    r2 = brainpylib_mv_prob_normal(events,
-                                   w_mu=w_mu,
-                                   w_sigma=w_sigma,
-                                   conn_prob=prob,
-                                   shape=(shape[1], shape[0]),
-                                   seed=seed,
-                                   outdim_parallel=outdim_parallel,
-                                   transpose=not transpose)
+    r2 = taichi_mv_prob_normal(events,
+                                          w_mu=w_mu,
+                                          w_sigma=w_sigma,
+                                          conn_prob=prob,
+                                          shape=(shape[1], shape[0]),
+                                          seed=seed,
+                                          outdim_parallel=outdim_parallel,
+                                          transpose=not transpose)
     c = jnp.allclose(r1, r2)
     if not c:
       print(r1, r2)
@@ -476,19 +465,20 @@ class Test_matvec_prob_conn(parameterized.TestCase):
     rng = bm.random.RandomState()
     events = bm.as_jax(rng.random((10, shape[0] if transpose else shape[1])))
 
-    f1 = jax.vmap(lambda e: brainpylib_mv_prob_normal(e,
-                                                      w_mu=0.,
-                                                      w_sigma=1.,
-                                                      conn_prob=prob,
-                                                      shape=shape,
-                                                      seed=seed,
-                                                      outdim_parallel=outdim_parallel,
-                                                      transpose=transpose))
+    f1 = jax.vmap(lambda e: taichi_mv_prob_normal(e,
+                                                             w_mu=0.,
+                                                             w_sigma=1.,
+                                                             conn_prob=prob,
+                                                             shape=shape,
+                                                             seed=seed,
+                                                             outdim_parallel=outdim_parallel,
+                                                             transpose=transpose))
     r1 = f1(events)
     r2 = f1(events)
-    c = jnp.allclose(r1, r2)
+    c = jnp.allclose(r1, r2, atol=1e-6)
     if not c:
       print(r1, r2)
+      print(r1 - r2)
     self.assertTrue(c)
 
     if x64:
@@ -528,7 +518,7 @@ class Test_matvec_prob_conn(parameterized.TestCase):
     events = events.astype(float)
 
     f1 = jax.grad(
-      lambda e, w_sigma: brainpylib_mv_prob_normal(
+      lambda e, w_sigma: taichi_mv_prob_normal(
         e,
         w_mu=0.,
         w_sigma=w_sigma,
@@ -537,12 +527,10 @@ class Test_matvec_prob_conn(parameterized.TestCase):
         seed=seed,
         outdim_parallel=outdim_parallel,
         transpose=transpose
-      ).sum()
+      )[0].sum()
     )
     r1 = f1(events, 1.)
     r2 = f1(events, 2.)
-    print('r1:', r1)
-    print('r2:', r2)
     self.assertTrue(bm.allclose(r1 * 2., r2))
 
     if x64:
